norme ok add strcmp.c

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: fcatinau <fcatinau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/09 15:32:53 by fcatinau          #+#    #+#             */
-/*   Updated: 2021/06/25 15:29:54 by fcatinau         ###   ########.fr       */
+/*   Updated: 2021/05/26 17:10:50 by fcatinau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_atoi(char *c)
+int	ft_atoi(char *c)
 {
-	int		i;
-	int		sign;
-	long	total;
+	int	i;
+	int	sign;
+	int	total;
 
 	total = 0;
 	i = 0;
